Extract lookup_class to reusable helpers module.
